+++ dify_studio/kaye_chat/nodes/gen_meta_content.py
@@ -1,4 +1,6 @@
 # pylint: disable=missing-module-docstring
+# pylint: disable=too-many-arguments
+# pylint: disable=too-many-positional-arguments
 
 # bug s/X replied 2 meta content during 2nd round conversation