+++ dify_studio/kaye_chat/nodes/post_start.py
@@ -6,11 +6,39 @@ OUTPUT_ROLE_KEY = "role"
 OUTPUT_SKIP_KEY = "skip_sense"
 
 
+# helpers  #####################################################################
+def should_skip_sense(role, llm_override):
+    """
+    decide whether should skip sense node for this round of conversation
+    based on combination of difficulty & LLM overrides given
+
+
+    :param role:
+    :type role: str
+    :param llm_override:
+    :type llm_override: str
+    :return: whether should skip sense node
+    :rtype: bool
+    """
+    if role == "barista":
+        # skip for Assistant Barista
+        return True
+
+    elif role == "changelog":
+        # skip for Changelog writer
+        return True
+
+    elif role:
+        # both both & LLM are provided
+        return bool(llm_override)
+
+    return False
+
+
 # Entry Point  #################################################################
 def main(
     role_override: str,
     llm_override: str,
-    difficulty_override: float,
     current_role: str,
 ):
     """
@@ -22,8 +50,6 @@ def main(
     :type role_override: str
     :param llm_override:
     :type llm_override: str
-    :param difficulty_override:
-    :type difficulty_override: float
     :param current_role: role saved in conversation variable,
             often decided on 1st round of conversation
     :type current_role: str
@@ -35,10 +61,6 @@ def main(
     # decide role  =============================================================
     role = role_override or current_role or ""
 
-    # decide if skip sense  ====================================================
-    skip_pre_sense = (
-        role == "coder" and (difficulty_override or llm_override)
-    ) or (role and llm_override)
-    # Bug dont skip for coder, b/c still need to extract PLs during sensing
+    skip_sense = should_skip_sense(role, llm_override)
 
-    return {OUTPUT_ROLE_KEY: role, OUTPUT_SKIP_KEY: bool(skip_pre_sense)}
+    return {OUTPUT_ROLE_KEY: role, OUTPUT_SKIP_KEY: skip_sense}